rename and move window class
