+++ client/mainmenu/CMainMenu.cpp
@@ -309,11 +309,11 @@ void CMainMenu::playIntroVideos(bool playVideoIntro)
 	{
 		auto playVideo = [](std::string video, bool rim, float scaleFactor, std::function<void(bool)> cb){
 			if(CCS->videoh->open(VideoPath::builtin(video), scaleFactor))
-				GH.windows().createAndPushWindow<CampaignRimVideo>(VideoPath::builtin(video), rim ? ImagePath::builtin("INTRORIM") : ImagePath::builtin(""), true, scaleFactor, [cb](bool skipped){ cb(skipped); });
+				GH.windows().createAndPushWindow<VideoWindow>(VideoPath::builtin(video), rim ? ImagePath::builtin("INTRORIM") : ImagePath::builtin(""), true, scaleFactor, [cb](bool skipped){ cb(skipped); });
 			else
 				cb(true);
 		};
-		playVideo("3DOLOGO.SMK", false, 1, [playVideo, playMusic](bool skipped){
+		playVideo("3DOLOGO.SMK", false, 1.25, [playVideo, playMusic](bool skipped){
 			if(!skipped)
 				playVideo("NWCLOGO.SMK", false, 2, [playVideo, playMusic](bool skipped){
 					if(!skipped)
